rpc/server: rework submitblock empty test
The original intention here was to check for server reaction in presence of
validation errors (response.ErrValidationFailed) and it was very easy to make
validation fail for block without transactions in Neo 2. But in Neo 3
transactionless blocks are perfectly valid (see
29d321b5e11d2e1f8a88ff169e2c56830e88fba9) which broke this test even though we
didn't see it until websocket addition (which required
8cec6694ae9e90f7b279fa5b5d66d5bf07fb9627 to make this test work initially).

So make a valid block and then spoil its verification script.