docker: drop default privnet-blocks bootstraps
They're no longer useful, their primary goal was to mint some GAS, but we have
plenty of GAS now right in the genesis block. They also contained NEO transfer
to non-multisig account, but that's easy to replicate using CLI. So keep the
bootstrapping feature (just in case), but drop the default files used for it.